requirements.txt
numpy==1.23.5
streamlit==1.36.0
opencv-python==4.10.0.84
opencv-python-headless
opencv_contrib_python
pandas==1.5.3
ultralytics==8.2.100
pillow==9.4.0